requirements.txt
aiopg==1.4.0
async-timeout==4.0.2
asyncio==3.4.3
get-docker-secret==2.0.0
iniconfig==2.0.0
loguru==0.7.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
